AGENT-183: Use gremlin.apparmor only when supplied by user input
A recent change broke the Gremlin Helm Chart for systems that did not
have apparmor. This change modify's this chart's behavior such that the
new apparmor annotation is applied only when the user supplies `--set
gremlin.apparmor=...`.

While this means we don't set apparmor to the exact profile Gremlin
desires, we allow it to be configurable enough to work for systems with
and without apparmor.

In a later change, I propose we do the same for SELinux, and
automatically set proper values based on some discovery of what the
system supports. We'll start with user input first.