rename pullSecrets to pullSecret for accuracy
Since we support just one secret (as opposed to the list accepted by `imagePullSecrets`), it makes sense to clarify this restriction in the variable name. Also, if we decide to support passing multiple secrets in the future, we leave ourselves open to using `pullSecrets` when that time comes, without producing a breaking change.