[Gremlin] Add Namespace to Secret template
**Background**
When you run the `helm install` command you can supply a namespace which
populates this field.  It also causes these templates to be inserted
into that namespace.  However, if you run other commands (eg. helm
template) then the namespace field only get populated on those templates
which make explicit references to it

It looks like secret.yaml is the only one which does not have this field
set

**Change**
Add this field